+++ harmony_browse_image_generator/utilities.py
@@ -1,4 +1,5 @@
 """ Module containing utility functionality. """
+import re
 from mimetypes import guess_type as guess_mime_type
 from os.path import splitext
 from pathlib import Path
@@ -10,6 +11,34 @@ KNOWN_MIME_TYPES = {
     '.jgw': 'text/plain',
     '.pgw': 'text/plain'
 }
+
+
+def get_tiled_file_extension(file_name: Path) -> str:
+    """Return the correct extention to add to a staged file.
+
+    Harmony's generate output filename can drop an extention incorrectly, so we
+    generate the correct one to pass in.
+
+    """
+    ext_pattern = r"(\.r\d+c\d+)?\.(png|jpg|pgw|jgw|txt)(.aux.xml)?"
+    match = re.search(ext_pattern, file_name.name)
+    return match.group()
+
+
+def get_asset_name(name: str, url: str) -> str:
+    """Return the name of the asset.
+
+    For tiled assets, we need to create a unique name beyond just the moniker
+    of data, metadata, or auxiliary in order to store each item in the asset
+    dictionary.
+
+    """
+
+    tiled_pattern = r"\.(r\d+c\d+)\."
+    tile_id = re.search(tiled_pattern, url)
+    if tile_id is not None:
+        name = f'{name}_{tile_id.groups()[0]}'
+    return name
 
 
 def get_file_mime_type(file_name: Path | str) -> str | None: